BOAC-4046, 'copy template' and 'create per student' use same clone-template code
